genai_local_chatbot/requirements.txt
flask
faiss-cpu
sentence-transformers
torch
streamlit